Move SerpAPI to utilities; add raw results method
There's a lot of valuable data that can come from the serpapi, more than just a single snippet or answer.

Add a `results` method which returns the raw api response from serpapi
