+++ langchain/utilities/serpapi.py
@@ -26,46 +26,6 @@ class HiddenPrints:
         sys.stdout = self._original_stdout
 
 
-def _get_default_params() -> dict:
-    return {
-        "engine": "google",
-        "google_domain": "google.com",
-        "gl": "us",
-        "hl": "en",
-    }
-
-
-def process_response(res: dict) -> str:
-    """Process response from SerpAPI."""
-    if "error" in res.keys():
-        raise ValueError(f"Got error from SerpAPI: {res['error']}")
-    if "answer_box" in res.keys() and "answer" in res["answer_box"].keys():
-        toret = res["answer_box"]["answer"]
-    elif "answer_box" in res.keys() and "snippet" in res["answer_box"].keys():
-        toret = res["answer_box"]["snippet"]
-    elif (
-        "answer_box" in res.keys()
-        and "snippet_highlighted_words" in res["answer_box"].keys()
-    ):
-        toret = res["answer_box"]["snippet_highlighted_words"][0]
-    elif (
-        "sports_results" in res.keys()
-        and "game_spotlight" in res["sports_results"].keys()
-    ):
-        toret = res["sports_results"]["game_spotlight"]
-    elif (
-        "knowledge_graph" in res.keys()
-        and "description" in res["knowledge_graph"].keys()
-    ):
-        toret = res["knowledge_graph"]["description"]
-    elif "snippet" in res["organic_results"][0].keys():
-        toret = res["organic_results"][0]["snippet"]
-
-    else:
-        toret = "No good search result found"
-    return toret
-
-
 class SerpAPIWrapper(BaseModel):
     """Wrapper around SerpAPI.
 
@@ -81,7 +41,14 @@ class SerpAPIWrapper(BaseModel):
     """
 
     search_engine: Any  #: :meta private:
-    params: dict = Field(default_factory=_get_default_params)
+    params: dict = Field(
+        default={
+            "engine": "google",
+            "google_domain": "google.com",
+            "gl": "us",
+            "hl": "en",
+        }
+    )
     serpapi_api_key: Optional[str] = None
     aiosession: Optional[aiohttp.ClientSession] = None
 
@@ -130,15 +97,19 @@ class SerpAPIWrapper(BaseModel):
             async with self.aiosession.get(url, params=params) as response:
                 res = await response.json()
 
-        return process_response(res)
+        return self._process_response(res)
 
     def run(self, query: str) -> str:
         """Run query through SerpAPI and parse result."""
+        return self._process_response(self.results(query))
+
+    def results(self, query: str) -> dict:
+        """Run query through SerpAPI and return the raw result"""
         params = self.get_params(query)
         with HiddenPrints():
             search = self.search_engine(params)
             res = search.get_dict()
-        return process_response(res)
+        return res
 
     def get_params(self, query: str) -> Dict[str, str]:
         """Get parameters for SerpAPI."""
@@ -149,7 +120,34 @@ class SerpAPIWrapper(BaseModel):
         params = {**self.params, **_params}
         return params
 
+    @staticmethod
+    def _process_response(res: dict) -> str:
+        """Process response from SerpAPI."""
 
-# For backwards compatibility
+        if "error" in res.keys():
+            raise ValueError(f"Got error from SerpAPI: {res['error']}")
+        if "answer_box" in res.keys() and "answer" in res["answer_box"].keys():
+            toret = res["answer_box"]["answer"]
+        elif "answer_box" in res.keys() and "snippet" in res["answer_box"].keys():
+            toret = res["answer_box"]["snippet"]
+        elif (
+            "answer_box" in res.keys()
+            and "snippet_highlighted_words" in res["answer_box"].keys()
+        ):
+            toret = res["answer_box"]["snippet_highlighted_words"][0]
+        elif (
+            "sports_results" in res.keys()
+            and "game_spotlight" in res["sports_results"].keys()
+        ):
+            toret = res["sports_results"]["game_spotlight"]
+        elif (
+            "knowledge_graph" in res.keys()
+            and "description" in res["knowledge_graph"].keys()
+        ):
+            toret = res["knowledge_graph"]["description"]
+        elif "snippet" in res["organic_results"][0].keys():
+            toret = res["organic_results"][0]["snippet"]
 
-SerpAPIChain = SerpAPIWrapper
+        else:
+            toret = "No good search result found"
+        return toret
